revert: bring back the client types which combine Query+Store
