+++ index.js
@@ -1,13 +1,3 @@
-const Endpoint = require('./Endpoint')
-const BaseClient = require('./BaseClient')
 const StreamClient = require('./StreamClient')
-const SimpleClient = require('./SimpleClient')
-const StreamStore = require('./StreamStore')
 
-module.exports = {
-  Endpoint,
-  BaseClient,
-  StreamClient,
-  SimpleClient,
-  StreamStore
-}
+module.exports = StreamClient
